Handle location services denial
In order to do this, I had to do a sizeable refactor of our permissions
requesting code. This involved hoisting a bunch of values to the class-
level to avoid excessive passing, including the button (which is now
shown and hidden based on permissions) the map, and the activity.

In the event that there are no services, the map auto-centers on
Cheesman Park