Separate TOU code out into TermsOfUseManager
This squirrels away all of the business logic RE Terms of Use into its
own object. The only logic left in MainActivity is view logic: whether
or not to show the modal and what to do when user interactions (in this
case, accepting.) This serves us in two ways:

(a) better organization. Yay!

(b) this will make for an easier refactor when we separate our the
    Terms of Use flow into its own activity