Move TermsOfUse flow into separate activity
This is nice for a bunch of reasons: all of the code (including the view
components) are now separated out from the main activity. Now the only
thing that MainActivity.kt knows about TermsOfUse is that on boot it
should call an activity by that name. Everything else handled there.