Add About page with map legend
I tried hard to turn this into a programmatically-generated map fragment
but ultimately couldn't figure it out, so I left a lot of the styling in
the view. I'm sure that this is a bad practice, but at the end of the day
you can only sink so many hours into doing this "the right way" before
you've gotta throw in the towel and just do it the easy way.

Where possible, I used @dimen and @color references so that if things
change those changes will be reflected in the legend.

Next up comes the settings.