Allow the user to toggle between street and satellite views
This was surprisingly hard, but I got it! Same as the other preferences,
the style will now be read from the user's SharedPreference, which they
selected on the settings/about page.

One odd thing to note here: there is what would appear to be some
duplicate code in setupMapBox: the setStyle method is called
identically in both clauses of the same if statement. Here's the
weirdness: they're not the same. setStyle is an overloaded method that
can take either a string (like Style.SATELLITE) or a Builder object,
like the one that's produced by our custom JSON file.

Kotlin isn't smart enough to figure out which one to route to, so I
have to invoke both.